requirements.txt
# Flask и расширения
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.1
Flask-WTF>=1.2.1
WTForms>=3.1.0
email-validator>=2.0.0

# База данных
SQLAlchemy>=2.0.23

# Работа с окружением
python-dotenv>=1.0.0

# AI и распознавание речи
assemblyai>=0.17.0
torch>=2.1.0
torchaudio>=2.1.0

# Обработка видео и аудио
ffmpeg-python>=0.2.0
librosa>=0.10.1
soundfile>=0.12.1

# Визуализация
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.24.0

# HTTP запросы и работа с API
requests>=2.31.0
urllib3>=2.1.0
certifi>=2023.11.17

# GigaChat SDK
gigachat>=0.1.40

# Markdown для форматирования текста
markdown>=3.5.0

# Дополнительные утилиты
Pillow>=10.1.0
scipy>=1.11.0

# Обработка аудио (зависимости librosa)
numba>=0.58.0
decorator>=5.1.0
audioread>=3.0.0
pooch>=1.8.0
soxr>=0.3.0
lazy-loader>=0.3.0
msgpack>=1.0.0
joblib>=1.3.0
threadpoolctl>=3.2.0

# Для работы с JSON и данными
simplejson>=3.19.0

# Дополнительные зависимости Flask
click>=8.1.0
itsdangerous>=2.1.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
blinker>=1.7.0

# Для работы с файлами
pathlib>=1.0.0
